+++ src/api/comments.request.ts
@@ -1,0 +1,1 @@
+import { CRUDBaseClass } from "./CRUDRequestClass";
